backend/requirements.txt
fastapi
uvicorn[standard]
pymupdf
pdfplumber
opencv-python
pillow
easyocr
langchain
langchain-community
faiss-cpu
sentence-transformers
google-generativeai
python-multipart
numpy
scipy
python-dateutil
langdetect
scikit-image
fuzzywuzzy
python-Levenshtein
